+++ lib/elixir_lokalise_api/collections/segments.ex
@@ -1,0 +1,12 @@
+defmodule ElixirLokaliseApi.Collection.Segments do
+  @moduledoc false
+  defstruct items: [],
+            project_id: nil,
+            total_count: nil,
+            page_count: nil,
+            per_page_limit: nil,
+            current_page: nil,
+            branch: nil,
+            language_iso: nil,
+            key_id: nil
+end